Tweaks

+++ sample-mfg/simulate-mfg.sh
@@ -11,14 +11,17 @@
 usage() {
     exitCode=${1:-0}
     cat << EndOfMessage
-Usage: ${0##*/} <priv-key-file> [<customer-pub-key-file>]
+Usage: ${0##*/} [<mfg-priv-key-file>] [<owner-pub-key-file>]
 
 Arguments:
-  <priv-key-file>  Device manufacturer private key?
-  <owner-pub-key-file>  Device customer/owner public key. If not specified, it will use a sample public key.
+  <mfg-priv-key-file>  Device manufacturer private key. If not specified, it will use SDO-support/sample-mfg/keys/sample-mfg-key.p12 (only valid for dev/test/demo)
+  <owner-pub-key-file>  Device customer/owner public key. This is needed to extend the voucher to the owner. If not specified, it will use SDO-support/keys/sample-owner-key.pub (only valid for dev/test/demo)
 
-Environment Variables that must be set:
-  SDO_RV_DEV_IP (will no longer be required when using the real RV service)
+Required Environment Variables:
+  SDO_RV_URL: usually the dev RV running in the sdo-owner-services. To use the real Intel RV service, set to http://sdo-sbx.trustedservices.intel.com or http://sdo.trustedservices.intel.com and register your public key with Intel.
+
+Optional Environment Variables:
+  SDO_SAMPLE_MFG_KEEP_SVCS: set to 'true' to skip shutting down the mfg docker containers. This is faster if running this script repeatedly.
 
 ${0##*/} must be run in a directory where it has access to create a few files and directories.
 EndOfMessage
@@ -27,28 +30,16 @@ EndOfMessage
 
 if [[ "$1" == "-h" || "$1" == "--help" ]]; then
     usage 0
-elif [[ -z "$1" ]]; then
-    usage 1
 fi
-: ${SDO_RV_DEV_IP:?}
+: ${SDO_RV_URL:?}
 
-privateKeyFile="$1"   # not yet sure what this is
-rvIp="$SDO_RV_DEV_IP"   #todo: the mfg won't have to specify this when using the real/intel RV
 sampleMfgRepo=${SDO_SAMPLE_MFG_REPO:-https://raw.githubusercontent.com/open-horizon/SDO-support/master}
-ownerPubKeyFile=${2:-$sampleMfgRepo/sample-mfg/owner-key.pub}
+privateKeyFile=${1:-$sampleMfgRepo/sample-mfg/keys/sample-mfg-key.p12}
+ownerPubKeyFile=${2:-$sampleMfgRepo/keys/sample-owner-key.pub}
+rvUrl="$SDO_RV_URL"
 
 dbUser='sdo_admin'
 dbPw='sdo'
-
-if [[ ! -f $privateKeyFile ]]; then
-    echo "Error: $privateKeyFile does not exist"
-    exit 1
-fi
-
-if [[ ${ownerPubKeyFile:0:4} != 'http' && ! -f $ownerPubKeyFile ]]; then
-    echo "Error: $ownerPubKeyFile does not exist"
-    exit 1
-fi
 
 # Only echo this if VERBOSE is 1 or true
 verbose() {
@@ -62,10 +53,24 @@ chk() {
     local exitCode=$1
     local task=$2
     local dontExit=$3   # set to 'continue' to not exit for this error
-    if [[ $1 == 0 ]]; then return; fi
+    if [[ $exitCode == 0 ]]; then return; fi
     echo "Error: exit code $exitCode from: $task"
     if [[ $dontExit != 'continue' ]]; then
         exit $exitCode
+    fi
+}
+
+# Check both the exit code and http code passed in and exit if non-zero
+chkHttp() {
+    local exitCode=$1
+    local httpCode=$2
+    local task=$3
+    local dontExit=$4   # set to 'continue' to not exit for this error
+    chk $exitCode $task
+    if [[ $httpCode == 200 ]]; then return; fi
+    echo "Error: http code $httpCode from: $task"
+    if [[ $dontExit != 'continue' ]]; then
+        exit $httpCode
     fi
 }
 
@@ -88,63 +93,89 @@ parseVoucher() {
     echo "${uuid:0:8}-${uuid:8:4}-${uuid:12:4}-${uuid:16:4}-${uuid:20}"
 }
 
-confirmcmds grep curl docker docker-compose java
+confirmcmds grep curl ping docker docker-compose java
 
-# Define the hostname used to find the SCT services (only if its not already set)
-echo "Adding SCT to /etc/hosts, if not there ..."
-grep -qxF '127.0.0.1 SCT' /etc/hosts || sudo sh -c "echo '127.0.0.1 SCT' >> /etc/hosts"
-chk $? 'adding SCT to /etc/hosts'
-
-#todo: remove this when it is time to use the real RV and OwnerSDO
-echo "Adding '$rvIp RVSDO OwnerSDO' to /etc/hosts, if not there ..."
-if grep -q RVSDO /etc/hosts; then
-    # In case we are using a different RV IP than before
-    [ $(uname) == "Darwin" ] || sudo sed -i -e "s/^.\+ RVSDO.*$/$rvIp RVSDO OwnerSDO/" /etc/hosts
-else
-    sudo sh -c "echo \"$rvIp RVSDO OwnerSDO\" >> /etc/hosts"
+# Initial checking of input
+# The mfg private key is either a URL we retrieve, or a file we use as-is
+mkdir -p keys
+if [[ ${privateKeyFile:0:4} == 'http' ]]; then
+    echo "Getting $privateKeyFile ..."
+    httpCode=$(curl -w "%{http_code}" --progress-bar -o keys/sdo.p12 $privateKeyFile)
+    chkHttp $? $httpCode 'getting mfg private key'
+    privateKeyFile='keys/sdo.p12'
+elif [[ ! -f $privateKeyFile ]]; then
+    echo "Error: $privateKeyFile does not exist"
+    exit 1
 fi
-chk $? 'adding RVSDO OwnerSDO to /etc/hosts'
-
-# Get the other files we need from our git repo
-echo "Getting docker files from $sampleMfgRepo ..."
-#set -x
-curl --progress-bar -o Dockerfile-mariadb $sampleMfgRepo/sample-mfg/Dockerfile-mariadb
-chk $? 'getting sample-mfg/Dockerfile-mariadb'
-curl --progress-bar -o Dockerfile-manufacturer $sampleMfgRepo/sample-mfg/Dockerfile-manufacturer
-chk $? 'getting sample-mfg/Dockerfile-manufacturer'
-curl --progress-bar -o docker-compose.yml $sampleMfgRepo/sample-mfg/docker-compose.yml
-chk $? 'getting sample-mfg/docker-compose.yml'
-# { set +x; } 2>/dev/null
 
 # The owner public key is either a URL we retrieve, or a file we use as-is
-mkdir -p keys
 if [[ ${ownerPubKeyFile:0:4} == 'http' ]]; then
     echo "Getting $ownerPubKeyFile ..."
-    curl --progress-bar -o keys/owner-key.pub $ownerPubKeyFile
-    chk $? 'getting owner public key'
+    httpCode=$(curl -w "%{http_code}" --progress-bar -o keys/owner-key.pub $ownerPubKeyFile)
+    chkHttp $? $httpCode 'getting owner public key'
     ownerPubKeyFile='keys/owner-key.pub'
+elif [[ ! -f $ownerPubKeyFile ]]; then
+    echo "Error: $ownerPubKeyFile does not exist"
+    exit 1
 fi
 
-# Copy the mfg private key to keys/sdo.p12, unless it is already there
-if [[ $privateKeyFile != 'keys/sdo.p12' || $privateKeyFile != './keys/sdo.p12' ]]; then
+# Ensure RV hostname is resolvable and pingable
+rvHost=${rvUrl#http*://}   # strip protocol
+rvHost=${rvHost%:*}   # strip optional port
+if ! ping -c 1 -w 5 $rvHost > /dev/null 2>&1 ; then
+    echo "Error: host $rvHost is not resolvable or pingable"
+    exit 1
+fi
+
+# If node is registered (if you have run this script before), then unregister it
+if which hzn >/dev/null; then
+    if [[ $(hzn node list 2>&1 | jq -r '.configstate.state' 2>&1) == 'configured' ]]; then
+        hzn unregister -f
+    fi
+fi
+
+# Get the other files we need from our git repo
+echo "Getting $sampleMfgRepo/sample-mfg/docker-compose.yml ..."
+#set -x
+httpCode=$(curl -w "%{http_code}" --progress-bar -o docker-compose.yml $sampleMfgRepo/sample-mfg/docker-compose.yml)
+chkHttp $? $httpCode 'getting sample-mfg/docker-compose.yml'
+# { set +x; } 2>/dev/null
+
+# Copy the mfg private key to the place docker-compose looks for it (keys/sdo.p12), unless it is already there
+if [[ $privateKeyFile != 'keys/sdo.p12' && $privateKeyFile != './keys/sdo.p12' ]]; then
     cp $privateKeyFile keys/sdo.p12
 fi
 
 # Start mfg services (originally done by SCT/startup-docker.sh)
-echo "Pulling and starting the SDO SCT services..."
+echo "Pulling and tagging the SDO SCT services..."
 docker pull openhorizon/manufacturer:latest
 docker tag openhorizon/manufacturer:latest manufacturer:latest
 docker pull openhorizon/sct_mariadb:latest
 docker tag openhorizon/sct_mariadb:latest sct_mariadb:latest
-# need to explicitly set the project name, because it was built under Services/SCT which by default sets the project name to SCT
+
+echo "starting the SDO SCT services (will take about 75 seconds)..."
+# need to explicitly set the project name, because it was built with that project name (see Makefile)
 docker-compose --project-name SCT up -d --no-build
 chk $? 'starting SDO SCT services'
 
-# Add the customer public key to the mariadb
+# sdo_sdk_binaries_linux_x64/SupplyChainTools/docker_manufacturer/mt_config.sql puts http://sdo-sbx.trustedservices.intel.com:80 in the mt_server_settings table
+# as the RV URL. Update that to the value the user wants to use.
+echo "Updating the RV hostname in the mt_server_settings table..."
+docker exec -t mariadb mysql -u$dbUser -p$dbPw -D intel_sdo -e "update mt_server_settings set rendezvous_info = '$rvUrl' where id = 1"
+chk $? 'updating RV hostname in mt_server_settings'
+
+# To enable re-running this script w/o shutting down the docker containers, we have to delete the voucher row from rt_ownership_voucher because it has a
+# foreign key to the 1 row in the rt_customer_public_key, which we will be replacing in the next step.
+echo "Removing all rows from the rt_ownership_voucher table to enable redo..."
+docker exec -t mariadb mysql -u$dbUser -p$dbPw -D intel_sdo -e "delete from rt_ownership_voucher"
+chk $? 'deleting rows from rt_ownership_voucher'
+
+# Add the customer/owner public key to the mariadb
 echo "Adding owner public key $ownerPubKeyFile to the SCT services..."
 docker exec -t mariadb mysql -u$dbUser -p$dbPw -D intel_sdo -e "call rt_add_customer_public_key('all','$(cat $ownerPubKeyFile)')"
 chk $? 'adding owner public key to SDO SCT services'
 # it can be listed with: docker exec -t mariadb mysql -u$dbUser -p$dbPw -D intel_sdo -e "select customer_descriptor from rt_customer_public_key"
+# all of the tables can be listed with: docker exec -t mariadb mysql -u$dbUser -p$dbPw -D intel_sdo -e "show tables"
 
 # Device initialization (and create the ownership voucher)
 echo "Running device initialization..."
@@ -164,6 +195,12 @@ deviceUuid=${deviceOcFile%.oc}
 echo "Device UUID: $deviceUuid"
 cd ../../..
 
+# At this point, the mariadb has content in these tables:
+#   mt_server_settings: 1 row with RV URL
+#   rt_customer_public_key: 1 row with all forms of customer/owner public key concatenated
+#   rt_ownership_voucher: 1 row with ownership voucher
+#   mt_device_state: 1 row of device info
+
 # Extend the voucher to the owner and save it in the current dir (orginally done by SCT/sct-docker.sh)
 echo "Extending the voucher to the owner..."
 devSerialNum=$(docker exec -t mariadb mysql -u$dbUser -p$dbPw -D intel_sdo --skip-column-names -s -e "select device_serial_no from rt_ownership_voucher where customer_public_key_id is NULL")
@@ -179,8 +216,8 @@ fi
 # this is what extends the voucher to the owner, because the db already has the owner public key
 docker exec -t mariadb mysql -u$dbUser -p$dbPw -D intel_sdo -e "call rt_assign_device_to_customer('$devSerialNum','all')"
 chk $? 'assign voucher to owner'
-echo "Device and owner data in the DB:"
-docker exec -t mariadb mysql -u$dbUser -p$dbPw -D intel_sdo -e "select rt_ownership_voucher.device_serial_no, rt_customer_public_key.customer_descriptor from rt_ownership_voucher inner join rt_customer_public_key on rt_customer_public_key.customer_public_key_id=rt_ownership_voucher.customer_public_key_id"
+printf "Device serial and owner in the DB: "
+docker exec -t mariadb mysql -u$dbUser -p$dbPw -D intel_sdo --skip-column-names -s -e "select rt_ownership_voucher.device_serial_no, rt_customer_public_key.customer_descriptor from rt_ownership_voucher inner join rt_customer_public_key on rt_customer_public_key.customer_public_key_id=rt_ownership_voucher.customer_public_key_id"
 
 # get the voucher from the db
 if [[ -f voucher.json ]]; then
@@ -203,9 +240,8 @@ if [[ "$deviceUuid" != "$voucherDevUuid" ]]; then
     echo "Error: the device uuid in creds/saved ($deviceUuid) does not equal the device uuid in the voucher ($voucherDevUuid)"
     exit 4
 fi
-echo "The extended ownership voucher is in voucher.json"
 
-# Note: originally to-docker.sh would at this point put the voucher in the ocs db, but our hzn-import-voucher does that later
+# Note: originally to-docker.sh would at this point put the voucher in the ocs db, but our hzn-voucher-import does that later
 
 # Switch the device into owner mode
 cd sdo_sdk_binaries_linux_x64/demo/device
@@ -217,8 +253,17 @@ chk $? 'switching device to owner mode'
 cd ../../..
 
 # Shutdown mfg services
-echo "Shutting down SDO SCT services..."
-docker-compose --project-name SCT down
-chk $? 'shutting down SDO SCT services'
+if [[ "$SDO_SAMPLE_MFG_KEEP_SVCS" == '1' || "$SDO_SAMPLE_MFG_KEEP_SVCS" == 'true' ]]; then
+    echo "Leaving SCT services running, because SDO_SAMPLE_MFG_KEEP_SVCS=$SDO_SAMPLE_MFG_KEEP_SVCS"
+else
+    echo "Shutting down SDO SCT services..."
+    docker-compose --project-name SCT down
+    chk $? 'shutting down SDO SCT services'
+fi
 
+echo '-------------------------------------------------'
+echo "Device UUID: $deviceUuid"
+echo '-------------------------------------------------'
+
+echo "The extended ownership voucher is in file: voucher.json"
 echo "Device manufacturing initialization complete."
